(layout) Prevent nested button toggle
The snippet above has been modified so that script does not look for
button elements nested inside of a the parent collapsed item.
Previously if a button with a collapse data attribute was nested inside
of another collapse element, the button element would toggle open when
it's parent element was opened, which is unwanted behavior.

An example of this is is on the Packages Page. When the "Version
History" collapse item was opened, the "Show" button in the package
communication notes was also opened.