(layout) Adjust package communication style
The package communication area was restyled with the redesign, however
it did display consistently in the different areas of the package
page where it might show up. For example, in the Version History table,
the package communication was not taking advantage of these styles. Same
goes for the communication area if a package has been rejected.

The JavaScript has been rewritten slightly to allow it to run in
multiple spots on the same page.

The Version History table has been modified to allow the
communication area to take up an entire row when the "Show" button is
clicked under "Notes". This makes the communication more legible, and
keeps the theme consistent. The "Show More" link has also been updated
to a button with proper meta tags, which increases accessibility.