(layout) Package Search Redesign
The search box from the package page has been moved to the top with a
blue background behind it. This makes the search standout more so users
can easily find it.

The CSS that was controlling the look and feel of the search results
has been taken out of packages.css and moved into the base file so that
it can be used site wide. This has also been converted to scss to
better handle updates if that time comes. The gulpfile.js has one added
line which prevents the purgecss plugin from deleting these styles.