(docs) Enable search in docs area
A search bar has been added to the left side navigation within the
documentation area. This search box will search the site and
documentation only, and will keep the results within the documentation
area instead of shooting off into the packages repository.

Additions were made within the Documentation controller, and a
DocumentationViewModel has been added to support this search function,
and make it work like it does in other parts of the website.