(layout) Add two search options to top nav
The search bar in the top navigation now has two search buttons. The
user has the option of searching the documentation area, or packages.
If the user simply presses enter, the default search when on the home
page will go to the packages area. If the user does this while in the
documentation area, only the docs will be searched.