(layout) Make collapsed elements navigatable
If a collapsed element exist on the page that matches the hash in the
url, then the collapsed item will be expanded and scrolled to on load.
When a user clicks on a collapsed element, the hash will be updated in
the URL to allow easy copy/paste. This functions very similarly to how
tabs are currently working.

Some cshtml has been updated to accomidate these changes
(DisplayPackage.cshtml and Pricing.cshtml), and also to
unwrap nested collapsed items that didn't need to be, resulting in much
cleaner markup.