(layout) Make typewriter JS compatible with IE
The javascript controlling the typewriter animation was not supported
on Internet Explorer. As a result, the typewriter animation and any JS
after that inside the same file was not working properly and
producing errors. The snippet has been rewritten to work on all
browsers, and a result lazy loading images and the global search now
works on IE.

A few styles were updated on the home page and search box to help with
IE display.