(layout) Update preloader to pure css animation
The preloader used on some pages throughout the website has been updated
to use pure css animations instead of requiring javascript. Even if the
user does not have javascript enabled, they will still be able to view
any page with the preloader. A javascript snippit still exist that will
remove the preloader and animation from the DOM after the page has
loaded, however this snippit is not required for the page to load
correctly.

The preloader itself has been updated to a bouncing chocolatey icon logo
instead of a bouncing box.

The preloader snippet itself has been moved into a partial since it is
used on multiple pages of the website.