(layout) Style changes to top alert
Minor style changes to the top alert to prevent overlapping of text and
the close button. This is a one time fix and will not need to be
changed every time the text is updated.

Background color was changed to red instead of green because of the
importance of the notification.