(GH-675) Ensure non-maintainer admin can restart cacher
If you are an admin, currently you would be able to restart the cacher,
however this should only be a moderation-level aspect. Other areas of
the packages page that deal with moderation don't allow mods to do
moderation of their own packages, thus this should also be limited to
admins that are not also a maintainer of a package.