(layout) Global search in top navigation
The logic has been modified slightly to allow the search box in the top
navigation to be used on every page. If a page has a search box
elsewhere on the page, such as in the docs or package list, the search
box in the nav will not be visible.

The logic on `_SearchBox.cshtml` has been updated to check for the user
Action instead of URL. This helps in cases where the url contains a `/`
at the end, and reduces the number of `if` statements needed.