(layout) Highlight Virus Scan results
Currently the Virus Scan results are hard to read and identify any
scans that had positive results. For this reasoning, a highlighting
system has been put in place. Based on the range of positive results,
the number will be highlighted a different color. There is a key at the
top of the Virus Scan result area now too, which explains the color
coding system.

The levels are 1-5, then 6-10, and then 10+