(layout) Make forms with Machines contain a number
Any form with the "Number of Machines" filed must now require a number
to be considered valid. For example a valid response would be: 5, 5
hundred, or < 5. An invalid response would be: five, five hundred, less
than five.

This is set up in the same way that the controller is checking to see
if a user has entered an email that is known to send spam.