(GH-732) Accommodate combined file style naming
Currently there is one file "How-To-Parse-PackageParameters-Argument.md"
that combines using "-" and camel casing in the name. In return, the url
is being formated in an undesirable way. A new snipet has been added
to the docs controller that will handle these types of file naming
scenerios.

For example, before the url for this would be
"how-to-parse-packageparameters-argument", but now it will be returned
as "how-to-parse-package-parameters-argument".