(layout) Autoplay videos inside modals
A slight change to the markup and js for videos inside modal popups has
been made. This will allow the video to autoplay when the user chooses
to open the modal instead of having to click again. Not all browsers
support autoplay however, so in this case the video will function as it
has in the past, and the user will have to click to start the video.