(GH-182) Hide comment/review instructions
The comment/review instructions are not necessary to be shown every
single time a user adds a new comment and it takes up screen space.
Now, the instructions will be shown on the page until the user clicks
to "collapse" the instructions. Once the instructions are collapsed, a
cookie is set that will collapse the instructions section on load until
the user clears their cache.