(GH-736) Show files with no content
With the refactoring of the Files section to use razor syntax instead
of javascript, one piece of content was wrapped inside of a
`(!string.IsNullOrWhiteSpace(file.FileContent))` that was not supposed
to be. This was causing files with no content to not be displayed. The
Files section has been adjusted slightly so that files with no content
will still be displayed.