(maint) Rename PackagesLayout to MainLayout
When the redesign was in process the new layout was initially only for
the packages area, which is why it is named "PackagesLayout". Now that
the layout it being used site wide, it needs to be renamed to
"MainLayout".

A few references to the Layout file were removed because they were not
not required to be included.