(layout) Update footer styles
While the footer area still looks the same, some styles that were being
used by a class name such as `bg-medium-dark` on the `footer` element,
have been moved to the main css file. This was done because the footer
is mentioned in multiple layouts, and wanted a better solution to keep
consistency if new layouts are added in the future.