(layout) Add new Newsletter Layout
The `_BottomNavigation.cshtml` files has been refractored a bit and
split up so that the Newsletter portion of it can be moved to it's own
seperate file. This was done to allow for a new layout,
`NewsetterLayout.cshtml` which features the newsletter signup in the
bottom navigation. The `MainLayout.cshtml` no longer has the newsletter
signup in the bottom navigation. Currently, only community pages are
using the `NewsletterLayout.cshtml` layout.