(GH-740) Fix avatars next to package status
Sometimes the avatars next to a package with a "waiting" status show up
without a profile attached to them, which then links off to a 404 page.
This happens when the last username to respond was "chocolatey-ops" or
if the user does not provide an email. In this case, now the avatar will
only show if the person who last responded has a valid email address.

For the "updated" and "responded" package status, a new helper has been
added that will allow to show the last person who responded or updated.
In this case that would always be the owner of the package, not the
reviewer like it was previously set.