(GH-738) Update SimpleMDE to EasyMDE
The library that is used on the packages page comments/review area has
shown to have bugs on Adroid devices. After a bit of research, there
are several known bugs on their GitHub for Adroid and others. Others
suggested that the upgrade to EasyMDE fixed all their issues. EasyMDE
is a fork of SimpleMDE and is being updated and maintained more
regularly than SimpleMDE. For these reasons EasyMDE has replaced
SimpleMDE. Users will not notice much of a difference, except no bugs
on Adroid!