(layout) Make moderation filter appear in url
Previously the moderation status filter was not appearing in the url
when unless the user was already in the moderation queue and then chose
another filter. This is not inline with how the other filters are
working. By adding the class "d-none" instead of removing it altogether
for the normal view, it will still be hidden, but the default value will
be submitted when the rest of the filters are.

This was mostly done to allow fixes for other issues that deal with
incorrect number counts when filtering and searching the moderation
queue.