(GH-743) Make moderation filter count correct
When going to the moderation queue, you will see the counts of
"Updated, Responded, Submitted, Waiting, Unreviewed" packages.
However, if you click on one of those to take you to a page with those
filters applied, the count at the top does not reflect the same number
shown on the first page.

Some updates have been made to both the ListPackackages.cshtml and the
PackagesController to make these counts correct. Changes to the
ListPackages.cshtml will carry over to help fix other issues as well.

Instead of using the "LastIndex" number for the total count (which means
last result for that page, so max would be 30), the appropriate
variable is used per package status filter.