(GH-731) Make moderation search results correct
When searching for anything in the moderation queue, the actual number
of found results was getting added to the unknown package count, which
made the number of results increase dramatically and also not be
correct. When searching in the moderation queue, the unknown package
count has been removed from the total hits found, making the search
results accurate again.