(GH-741) Allow mod status filtering with search
While in the moderation queue and also searching for a specific package,
the moderation status filter would not work. It would allow you to
change the dropdown, but the results would remain the same. A fix for
this has been made to allow searching by term inside a package status
type. For example, a user will now be able to search for "choco" in
all "Ready" moderation package status types.