(layout) Resource nav to individual pages
In an effort to reduce the number of `if` statements based off url, the
subnavigation needed to supply users with basic access to the different
resources (like docs, videos, blog) has been moved into a partial and
placed on each individual page that needs it. This will significantly
reduce the number of request and improve page speed. It is also easily
managaed since the navigation is in a partial.