(layout) Switch if statements based off Action
Wherever possible, if there was an `if` statement looking for a current
url, it has been switched to look for the current action instead. This
has reduced some of the logic needed on each page. When searching for
the current url, sometimes 2 `if` statements were needed to check to see
if the url contained a `/` at the end of it or not. That is no longer
needed when using the action instead.