(GH-746) rewrite stats query for perf
The current stats query pulls information in a lot of different ways,
and does multiple selects against the same table. That is now taking a
very long time to run, thus rewrite it for performance to do very few
scans against a table.

The estimated query plan shows the updated query is 11% of the total
run versus the old method.