(GH-750) Allow enabling/disabling of background jobs
It is now possible to run these background jobs as a SQL Server Agent
job.  As such, it should be possible to choose how these tasks are
completed.  Using new settings in the web.config file, which will
default to false, either register these Jobs with the background manager
or not.