(GH-752) Prevent double "Preview" in text editor
Inside of any text editor that takes advantage of the EasyMDE library,
an eye with the word "Preview" is shown. In cases where there are more
than one of these text editors on a page, the word "Preview" is being
appended twice.

The code for appending the word "Preview" has been slightly modified to
address this problem. Since the code for this is inside an `each()`
loop, the identifier needs to originate from `$(this)`, or else the
error above will occur.