(layout) Fix mobile button margins
In cases where there are two buttons that are meant to be displayed
side-by-side on mobile, we do not want to apply a top margin. The
`btn-preferences` button (which is located in the users package
preferences model) was being referenced incorrectly in the css, which
was causing it to have a top margin on mobile, which is unwanted. The
CSS has been corrected to allow the buttons to be side-by-side properly.