(GH-756) Make internal doc links contain hyphens
Some of the internal links inside the documentation files do not contain
hyphens in between words when they are supposed to. The links still
work how they are, but they are pointing to a different url than what
has already been established.

A new method has been added to handle adding these hyphens in between
words much similarly as file names are being handled. This ensures all
urls are consistent throughout the documentation area.

For example, if the link below was referenced within a documentation
file, it would be rendered as

https://chocolatey.org/docs/gettingstarted#what-is-chocolatey

Now with the fix (correct link):
https://chocolatey.org/docs/getting-started#what-is-chocolatey