(GH-761) Fix: Lucene updating one package causes duplicates
When a package has been updated, the index should also be updated.
However when all versions of a package are stored in the index, all
versions need to be reviewed for updates. That is done in the
UpdateIndex() method, so pass over to it for review and update and
return afterwards to reduce the chances of issues.