(GH-766) FindPackagesById uses index when all
If the search service contains all versions, then finding packages by
id should use that to return all results instead of caching a query
into the database. To allow for paging/continuation token to work
correctly, all results need to be gathered at the first request and
cached for the subsequent requests.

Add a TakeAllResults to the search filter to ensure that the index
returns all results that it has for a query instead of limiting it to
the the MaxPageSize, which is currently set at 40.