(GH-776) Fix broken link in Getting Started course
In the Getting Started course, Terminology module, there is a bad link
"Package- See What are Chocolatey Packages?". This link does not go
anywhere when clicked.

The link has been updated to point to the appropriate documentation
that explains what a Chocolatey Package is. This link will open in a
new tab, so that the user does not lose their place in the course.