(GH-772) Support backwards markdown compatibility
Previously, when the website was using MarkdownSharp to convert
markdown to cshtml, it would support the use of `###Header` instead of
the correct `### Header` (notice the space). When changing over to
Markdig, this was no longer supported, which meant some older packages
that were using the `###Header` syntax were not getting converted
correctly.

In this PR, a few fixes were put in place to find these types of header
patterns (a hash followed by no space and a character). Now, if one of
these patterns are found, it will replace it with the correct syntax
before Markdig takes over. This means when Markdig does its thing, it
will be converted correctly onto the page resulting in the desired
outcome (backwards compatibility).