(layout) Make logo clickable on MinimalLayout
When on the login, register, signup pages etc. there was no way to
get back to the home page. Now the top "Chocolatey" logo is clickable,
and will take the user back home. This is typical behavior for a logo to
be clickable and take them to the top main level page.