(GH-783) Fix broken link to moderator queue
Currently, if a user is to go to
https://chocolatey.org/packages?q=&moderatorQueue=true
they will be hit with a 404 page.

This is due to the moderationStatus filter being null. Now, if the
filter is null then it will be defaulted to "AllModerationStatuses"
which will show all the packages in moderation as expected. With this
in place, the link above now works as intended.