(GH-785) Ensure view model lengths
Some fields in the view model (email address mostly) needs to be
restricted to the size of the field size stored in the database. This
will ensure that the site doesn't error if something larger is entered,
that it will instead go back to the user and restrict the length.

There is also a fix on password length that can be entered to match the
size the lengths that can be stored. This aspect fixes a security
finding by Jhon Vik.