(GH-785) Choose insensitive search
When searching against code related sources, or by default, convert
everything to lower case to evaluate. When using the Search for
database (case-insensitive) searches, turn that off so that the
search queries can be more efficient.

Also, fix the search terms to determine if the query is for an id,
author, or tag search at the top level, then loop through and add items
into the criteria based on the search call - this allows multiple words
with spaces to still be evaluated as part of a specific search. Also it
ensures all the search is done in a similar manner as the Lucene Search
service performs it.