(GH-796) Fix one off doc links w/ incorect casing
The filename "CommandsSetapiKey" suggest that it should be interpreted as
"commands-setapi-key" on chocolatey.org. However, the link desired is
"commands-setapikey" since that is the command name.

The filename "HelpersGetPackageParameters" suggest that it should be
interpreted as "helpers-get-package-parameters" on chocolatey.org.
However, the link desired is "helpers-get-packageparameters".

For this reason a snippet has been added to manually change the
filenames so that they are parsed correctly on load.