(GH-796) Make second level invalid URLs redirect
In a case where a URL is a second level, for example `/docs/fake-url`
or `blog/fake-url`, the `fake-url` will be redirected back to the top
level home page. For this example `blog/fake-url` would be redirected
back to `blog`. This reduces the number of
`System.InvalidOperationException` errors and provides a better user
experience than getting an "Oh no we broke something" page.