(seo) Make each package have share description
Right now, if a package is shared, the description that would appear is
the general description for the whole website. Now, if a package is
shared, or viewed on Google, the package description will appear
instead. This is better for SEO and also more descriptive to the link
being viewed.