(GH-818) Reformat navigation and update JS
Since this secondary authentication navigation which contains the
Logoff button is cached on some pages, the decision has been made to
have JS handle all the showing/hiding of elements based off the
Authentication Cookie. Some JS was already in place for this, but
sometimes the mixing of JS and Razor syntax still caused the wrong
navigation items to be shown.

The `_AuthenticationSubNavigation.cshtml` has been updated to reflect
this and any Razor syntax has been removed from showing/hiding
elements, except one which is based off a url route, and not whether
the user is authenticated or not.

The javascript still follows the same general concept, but now a
specified `ul` will be shown if the user is logged in or whether the
user has logged off. This has reduced the amount of JS that is needed
also, which increases page load times.