(GH-823) Handle null cookie key values
In `CookieTempDataProvider.cs` there was nothing to protect against null
key values. Although the website appeared to be running fine, this was
causing `ArgumentNullException` errors to be thrown in Elmah. Parts of
the file `CookieTempDataProvider.cs` has been updated to the latests
code which includes protection against this and handles null values
appropriately. The `secure` flag was also added to the cookie as good
practice (and is also in the updated file).

Reference:
https://github.com/NuGet/NuGetGallery/blob/master/src/NuGetGallery/Infrastructure/CookieTempDataProvider.cs