(GH-674) Update package statuses on profile page
Some of the necessary package statues were not presented on the profile
page, which made the made package status shown here and the package
status shown on the actual package different. The appropriate package
statuses have been added to the profile page so that both the actual
package and the status on the profile page match and are correct.