(GH-827) Fix JS on Install and Verify Package page
Currently when going to the Install or Verify Package (step 2 when
uploading a package) page, there is a JS error of `'querySelectorAll'
of undefined`, which is being caused by a snippet of JS that is meant
to be run specifically on the Package Display page.

This error is causing the install steps area not to work on the Install
page, and the file that contains this code is actually not needed at
all on the Verify Package page.

The JS snippet that controls the install steps has been moved to its
own JS file, which allows the package-details.js to be removed entirely
from the 2 pages that do not need it. This cuts down on the amount of
unnecessary JS on these 2 pages. The `install-steps.js` file has been
added to the Package Display page and the Install page as those are the
2 pages that require it at this time.

In result, the file that is causing the initial error has been removed
and contained to the specific page that is needed on which is the
Package Details page.