(layout) Update template used on Install page
The Install page is using the NewsLetter layout, which includes a
newsletter signup form in the footer area. Since there is already a
newsletter signup form in Step 1 of install, this form at the bottom is
not needed.

The duplicate form is also causing errors in the DOM because duplicate
ID's are found.