(GH-831) Create Ajax call to handle sub navigation
Since the Authentication cookie has been set to secure, the process in
which the authentication sub navigation where users login/logoff needed
to be updated. JavaScript is unable to interact with secure cookies
like it was previously to show/hide the correct elements.

Now, an ajax call will be made when a user is on a page where the
authentication sub navigation is meant to be shown. The ajax call will
determine if the user is authenticated or not and send a response
back to the JavaScript function. If the call is a success and the
returned data equals `true`, then logoff navigation will be shown. If
the call is a successs and the returned data equals `false`, then the
login data will be shown. This is a similar process as before, now just
dealing with ajax instead of determining off a cookie.