(GH-830) Allow package related forms to submit
Currently, the Contact Admins, Contact Owners, and Report Abuse forms
do not submit, and instead trigger the LogOut action. This is due to
the recaptcha JavaScript that is on these pages. JavaScript was
previously being used on forms that needed recaptcha to add an ID
element to the form, however this method does not work when there is
more than one form on the page. The LogOff form for was mistakingly
getting the recaptcha ID, which in tern was triggering another function
to make the LogOff form submit (and logoff the user) whenever the
intended form with recaptcha was submitted.

Instead of using JavaScript to identify forms that need the recaptcha
ID, the ID has been added using razor syntax to the specified forms that
need it. While only 3 forms need this at this time, the syntax has been
changed site wide to keep consistency, and to prevent future problems.