(GH-778) Update installation for TLS 1.2
Update the installation scripts and docs for TLS 1.2. Add in the
necessary bits in the installation scripts to ensure that the
PowerShell client uses 3072 (or TLS 1.2) to talk to the chocolatey.org
website.

Update all the docs for installation to show setting the minimum TLS
level so that installation is most likely going to be successful as
long as the clients have the minimum versions of PowerShell / .NET
Framework.