(GH-835) Get username on authentication via ajax
The first page of the package area `/packages` is heavily cached. For
this reason, we need to get the username via ajax call in order for the
`profiles` link in the sub navigation to work. Now that we have the
username, JavaScript can inject it on load, so that the link will work
even if the page is heavily cached.

Method of retrieving the username in `PagesController.cs` is similar to
how it is implemented in `UsersController.cs`.