(GH-838) Prevent flicker of noscript message
Sometimes on load the noscript message in the authentication navigation
is flashed on the screen before the JavaScript has a chance to remove
it.

This line of code to remove the noscript message has been moved to be
performed `onload` and also to be removed before the preloader is
removed.

This is the primary reason for having the preloader. It give the DOM a
little extra time to perform tasks that need to happen before the page
is shown to the user. The preloader has been added to the authentication
navigation itself so that it will be utilized on these pages. The
preloader partial has been removed from pages that already had it on
there, and that also have the authentication navigation, to prevent
duplication.