(layout) Add Self-Service Anywhere Page/Resources
A solutions area has been added to the website, curretly with one
solution "Self-Service Anywhere". With the addition of this page, new
resources were added to the resource area to be able to do display at
the bottom of the webpage.

Minor update to the custom.js to support styling of embeded videos on a
page to accomidate an overlay image which dissapears when the video is
start. (otherwise the video is black).