(maint) Remove Top Alert Only on "X" Click
Currently, the top alert banner will be removed once a user clicks on
it by using a session storage item. Now, the top alert banner will stay
in place unless the user specifically clicks on the "X" in the right
corner to remove it.

Minor updates to the JS snippet controlling this feature was made. The
changes also eliminates the flash of red sometimes seen on page load if
a user has removed the alert.