(layout) Adjust top navigation dropdown height
Whenever the top alert banner is active inside the navigation, and the
user clicks to dismiss the alert, and then goes to use the dropdown top
menu, the menu appears too far below the navigation creating a gap.

One area in the JS has been updated to change the position of the top
dropdown so this issue does not occur. The position is now calculated
by the outer height of the entire header which will accommodate
regardless of the viewport or whether the top alert is active or not.