(resources) Add Feature Category to Resource Area
Initial changes needed to add a new resource type of "Feature" to the
list. This will have it's own page, and on that page the "Feature"
videos will be further broken down into either C4B Features or Open
Source Features.

The videos that are in a Feature series will be ordered by file name
instead of published date. This is intended so that the videos stay in
the specified order. If the order needs to change of the series videos,
the file names will need ot be adjusted. Each feature video follows the
file name structure of "04-xx" where "xx" would be the position in the
series.