(maint) Update Links to Product Pages
A recent change updated the routes to the Products page to include
another sub page products/chocolatey-for-business where in the URL
`productName` points the the page that should be displayed.

In order to show the main url of /products, `productName` should be
equal to "home" because that is the top level page. The url's to the
main /products page have been updated throughout the website.