(blog) Update Markdig Extensions Used
A few items were updated in the BlogController over how Markdig is
rendering markdown. These changes are inline with how the markdown is
being rendered in the documentation area currently. Support has been
added for emojis and also for custom containers, which allow classes to
be added to markdown items, almost eliminating the need for inline html.