(layout) Events Area Template
The events area has been modified to pull information for each event
from a markdown file. This allows more people to be able to add a
new event on their own. A README has been added in the root of the
`Website/Views/Events` folder that takes a user through the steps of
adding a new page.

The main events page has also been added. This is automatically
updated by the addition of a markdown file. Once an event date has
passed, the event will be moved from the "upcoming events" section to
a new "watch on-demand" section.

JS has been added also to the countdown clock for each webinar page
itself. Now, if the event date has passed, the countdown clock will be
removed, and any time sensitive text will be updated or removed as well.