(GH-613) Contact Lucene less out of band
The Lucene cache updates every 10 minutes. Letting things hang for 10
minutes instead of trying to explicitly update Lucene could actually
improve the way it works and result in less Out of Memory Exceptions,
like that are currently being seen. Adjust so that some aspects of
putting up a new package are not adjusting the cache.