(GH-798) Remove ability to upload nupkg
This is known to cause a number of problems, and the recommendation has
been for a long time for folks to use the choco push command.  Removing
the upload through the site encourages people to use the recommended
way.

The page structure has been left in place in case anyone has shortcuts
to the path, and information has been left on that page for the correct
course of action.