(GH-671) Remove Codeplex Links From User Profile
Removes the ability to add a Codeplex link when updating user account
information, and also removes links on the users profile.profileRepo

A fix was put in on `Profiles.cshtml` to prevent the showing of any
Codeplex links that already exist in the database. At anytime should
these links be removed from the database, this fix on line 53 can be
reverted.