(GH-852) Make Search Box Autofocus on Load
Upon load, if a search box is present on the page (not including top
navigation), an no other forms have autofocus applied, then the search
box will be autofocused. This mostly impacts the search box on the
packages and documentation pages.

This feature was previously deployed with the redesign, but somewhere in
the mix of updating it broke. A slight update to some JS and classes
has made it work again.