(GH-902) Add ability to exempt validation
Added the ability to exempt a package from the package-validator. This
only happens for the current package version.  The intention is that
validation is by based for a specific version to allow moderation to
continue, and appropriate work will be done on the next package version
to fix the issues that the validator brought up.