(GH-905) Clarify what virus scan results mean
There is some ambiguity, both in the code, and in the resulting legend,
about the bounds of the light, medium and dark virus scan levels.  The
code has been updated to make it clear what those ranges are, and they
now match up to what is displayed in the legend.