(GH-909) Don't copy unnecessary files from wiki
The choco wiki includes a couple of files, _Sidebar.md and _Footer.md
which are specifically used to render information when viewed in GitHub.
These files are not required in the chocolatey.org version of the docs.
Updated the copy script to not include bringing these over, and deleted
them from this project.