(GH-922) Show Content Inside Collapsed Text Editor
There is a slight bug in the Easymde JS library used that when a text
editor is hidden on load inside a collapsed or tabbed item, the content
is not displayed until the editor is clicked on. A workaround has been
put in place that will basically "refresh" the editor whenever the
hidden editor becomes visible. A new class must be applied to the
collapsible item of "text-editor-refresh" in order for this JS to
trigger. Support has not been added in for editors that appear inside
tabbed items since there is currently none on the website and no plan
to add any.

More information on this bug and fix here:
https://github.com/Ionaru/easy-markdown-editor/issues/208#issuecomment-645656131