(events) Include Option to Redirect Registration
On some events, we do not want users registering from our website, but
instead want to redirect traffic to anther website. For example, if we
are doing a joint webinar with Puppet, we may want to send users to
Puppets website to register instead of our own.

A new option has been added of "IncludeRegisterPage" which defaults to
"true", meaning that a registration page will be created and used on
our website. If this value is set to "false", the register link will
redirect to another website, and no "learn more" buttons will appear,
only "register" buttons.