(layout) Add Right Side Announcement Bar
An announcement bar has been added that will appear on the right hand
side of the screen while on desktop, and full screen while on mobile.
The announcement bar is slightly different on the package listing page
then on the rest of the website, as it is static and always visible. On
the rest of the website, there is a new button that is located on the
right hand side of the screen that will show/hide this announcement
area when clicked. While on mobile, this switches to a button that is
added to the top navigation bar.

The announcements are set to "shuffle" on reload, so that a user will
potentially see different announcements while navigating throughout the
website.