(GH-939) Add Notification Badge to Button
Sometimes the button to the right side notification bar is not very
noticeable. To draw a bit more attention to this feature, a little red
badge has been added to the button if the user has not yet clicked it.
The badge also displays the number of "unread" notifications, much like
the notifications you might see on an app when it needs updating. Once
the user clicks on the button, this badge will disappear and a cookie
is set.

The next time the notification bar is updated, two hidden fields have
been added that will also need to change in order for the cookie to be
reset. Instructions on what to do have been added to the document.