(GH-947) Break Long Titles in Grid View
While in the grid view mode on the packages page, package titles will
now break if the title is too long to fit in the box instead of busting
out of the box. This fixes alignment issues with both the title and the
image placement. A Bootstrap class has been added to support this
feature. See:

https://getbootstrap.com/docs/4.4/utilities/text/#word-break