(GH-954) Update C4B Videos in Resource Area
Updates the C4B videos in the resource are with what is reflected in
the YouTube playlist. A Package Synchronizer video has been added,
and the Chocolatey Central Management video has been updated to include
Deployments.

These updates have been made in the resource area and also to the
Chocolatey for Business page.