(GH-960) Make Carousel Prev/Next Buttons Function
The prev/next buttons on any carousel item currently do not transition
between slides. This was due to a recent addition of the announcement
flyout, where a carousel was added with the same ID that may be found
on other pages. When the same ID is on the page twice, it breaks the
carousel. The ID has been changed to be unique and not match any other
carousel used throughout the website. Now, the prev/next buttons work
as intended.