(GH-937 )Update to reCAPTCHA v3
In an effort to reduce the amount of spam, reCAPTCHA has been updated
from v2 to v3. If reCAPTCHA returns a score of less than 0.5, it is
determined that this is likely a bot, and the form will not be allowed
to submit and an error message will appear on the form. The same applies
if the action returned does not contain key words from that was
submitted.

The SpamValidationResponse output on each page has been restructered
slightly since there can now be more than one response listed and also
to allow links to be used inside the error message.

As per directed in the SpamValidationResponse error message, the Gitter
chat has been added to the home page.

Docs:
https://developers.google.com/recaptcha/docs/v3